backend/requirements.txt
fastapi
uvicorn
python-multipart

numpy
scikit-learn
pillow

tensorflow